Refactor UnicornScene component and hooks for improved structure
- Moved the UnicornScene component and related hooks to separate directories for better organization.
- Updated the UnicornScene component to utilize the new hooks structure, enhancing modularity and reusability.
- Introduced a new useUnicornStudioScript hook to manage script loading and error handling more effectively.
- Re-exported types and constants for better accessibility across modules.
