+++ src/next/hooks.ts
@@ -1,0 +1,20 @@
+import { useState, useCallback } from "react";
+
+export function useUnicornStudioScript() {
+  const [isLoaded, setIsLoaded] = useState(false);
+  const [error, setError] = useState<Error | null>(null);
+
+  const handleScriptLoad = useCallback(() => {
+    setIsLoaded(true);
+  }, []);
+
+  const handleScriptError = useCallback(() => {
+    setError(new Error("Failed to load UnicornStudio script"));
+  }, []);
+
+  return { isLoaded, error, handleScriptLoad, handleScriptError };
+}
+
+// Re-export shared hooks
+export { useUnicornScene } from "../shared/hooks";
+export type { UseUnicornSceneParams } from "../shared/hooks";
